Fail case for foreign composite key ordering.
It's odd situation, and it should not fail but,

When the foreign key name is like,

`model_db_signal_model_name_db_type_version_fkey`

but the column order is like,

```hcl
  foreign_key "model_db_signal_model_name_db_type_version_fkey" {
    columns     = [column.model_name, column.version, column.db_type]
    ref_columns = [table.model_db.column.model_name, table.model_db.column.version, table.model_db.column.db_type]
    on_update   = NO_ACTION
    on_delete   = NO_ACTION
  }

```

Then it is failing because foreign key name order is different with column order.

Failing means, even the DB is same, `atlas migrate diff` keep regenerate the migration script which does 1) delete existing foreign key, and 2) recreate foreign key with same composite key.

So, when u check this commit, it only changes the order of the key, but failing happens.

Signed-off-by: Hyukjoong Chuck Kim <wangmir@gmail.com>